# server/swagger.yaml
openapi: 3.0.0
info:
  title: Smart Inventory Optimizer API
  description: API documentation for the Smart Inventory Optimizer project
  version: 1.0.0
servers:
  - url: https://smart-inventory-optimizer.vercel.app/
    description: SIO API
  - url: "http://localhost:{port}"
    description: Local server
    variables:
      port:
        default: "3000"
paths:
...
  /api/storage/delete:
    delete:
      summary: Delete profile image
      description: Deletes the profile image for the authenticated business.
      operationId: deleteURL
      tags:
        - Storage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Object deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "object deleted successfully"
        "401":
          description: Unauthorized - Authentication token is missing or invalid
        "403":
          description: Forbidden - Access denied
        "500":
          description: Internal Server Error - Could not delete object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT